# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie Review API",
        "description": "API para plataforma de reseñas de películas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/genres": {
            "get": {
                "tags": [
                    "Géneros"
                ],
                "summary": "Listar todos los géneros",
                "operationId": "eb12635064fefadf7ff7e54900a73534",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de géneros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de géneros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "genre": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "actual": {
                                                    "type": "integer"
                                                },
                                                "pages": {
                                                    "type": "integer"
                                                },
                                                "index": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Géneros"
                ],
                "summary": "Crear nuevo género (Solo admin)",
                "operationId": "8596dc3b1780529a7ce51429eee5ed0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "genre_name"
                                ],
                                "properties": {
                                    "genre_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Acción"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Género creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "genre_id": {
                                            "type": "integer"
                                        },
                                        "genre_name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado - Se requieren permisos de administrador"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genres/{id}": {
            "get": {
                "tags": [
                    "Géneros"
                ],
                "summary": "Mostrar un género específico",
                "operationId": "3e2ac2805d9f25d984bab075133bbb96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "genre_id": {
                                            "type": "integer"
                                        },
                                        "genre_name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Género no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Géneros"
                ],
                "summary": "Actualizar un género",
                "operationId": "65d9a201f51320eb4147a75f91a8119b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "genre_name"
                                ],
                                "properties": {
                                    "genre_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Acción"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Género actualizado exitosamente"
                    },
                    "404": {
                        "description": "Género no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Géneros"
                ],
                "summary": "Eliminar un género (Solo admin)",
                "operationId": "4ddc050edb0bf18c202c611a028f457f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género eliminado exitosamente"
                    },
                    "403": {
                        "description": "Acceso denegado - Se requieren permisos de administrador"
                    },
                    "404": {
                        "description": "Género no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}